% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{add_scores}
\alias{add_scores}
\title{add_scores}
\usage{
add_scores(player_dice, num_dice, num_games, num_rolls, strategy,
  game_ticker, gamecard, scores_vector, final_score, multiplayer)
}
\arguments{
\item{player_dice}{a vector of the dice saved to a player's hand in a given game}

\item{num_dice}{the number of dice a player has left to roll}

\item{num_games}{the number of games a user wishes to run within the simulation}

\item{num_rolls}{the number of rolls a player has left to use}

\item{strategy}{a character vector containing p1 and p2 strategies}

\item{game_ticker}{the number of games already completed in the simulation}

\item{gamecard}{a dataframe storing game number and score information}

\item{scores_vector}{a vector of a player's scores}

\item{final_score}{a numeric vector of a player's most recent final score}

\item{multiplayer}{a logical vector determining whether games are single or multiplayer}
}
\description{
Checks if all requested games have been played, if the games are single or multiplayer, and assigns the final gamecard to the global environment
}
\seealso{
Other helper functions: \code{\link{final_score_tab}},
  \code{\link{greedy_reroll}},
  \code{\link{initial_gameplay}},
  \code{\link{rolling_function}},
  \code{\link{strategy_function}}
}
\concept{helper functions}
