% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_helpers.R
\name{check_roll}
\alias{check_roll}
\title{check_roll}
\usage{
check_roll(rolled_dice, prev_dice)
}
\arguments{
\item{rolled_dice}{\itemize{
\item most recently rolled dice
}}

\item{prev_dice}{\itemize{
\item any dice kept from previous turns
}}
}
\value{
List containing a vector of the dice to be kept and a vector of dice to be rerolled
}
\description{
Helper function for interactive gameplay
Separate rolled dice into those that can be kept and those that can be rerolled
}
