% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_helpers.R
\name{find_score}
\alias{find_score}
\title{find_score}
\usage{
find_score(needed, cargo)
}
\arguments{
\item{needed}{vector to check if score will be nonzero}

\item{cargo}{vector of numbers used to calculate score}
}
\value{
number denoting player score
}
\description{
Calculate current player score
}
\seealso{
Other Interactive Gameplay: \code{\link{game_turn}},
  \code{\link{play_again}}, \code{\link{player_message}},
  \code{\link{reroll_prompt}}, \code{\link{reroll}},
  \code{\link{start_game}}
}
\concept{Interactive Gameplay}
