% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_functions.R
\name{game_turn}
\alias{game_turn}
\title{game_turn}
\usage{
game_turn(prev_rolls, dice_to_roll, prev_dice)
}
\arguments{
\item{prev_rolls}{Number of turns already completed}

\item{dice_to_roll}{Number of dice to be rolled this turn}

\item{prev_dice}{Any dice kept from previous turns}
}
\value{
List containing number of dice that can be rolled next turn, current player score, and dice that are kept
}
\description{
Simulate single game turn
}
\examples{
game_turn(0, 5, NULL)
game_turn(1, 3, c(6, 5))
}
\seealso{
Other Interactive Gameplay: \code{\link{find_score}},
  \code{\link{play_again}}, \code{\link{player_message}},
  \code{\link{reroll_prompt}}, \code{\link{reroll}},
  \code{\link{start_game}}
}
\concept{Interactive Gameplay}
