% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{multiplayer_game}
\alias{multiplayer_game}
\title{multiplayer_game}
\usage{
multiplayer_game(num_games, strategy, multiplayer = TRUE)
}
\arguments{
\item{num_games}{the number of games a user wishes to run within the simulation}

\item{strategy}{a character vector containing p1 and p2 strategies}
}
\description{
Initializes gameplay for a two player simulation of Ship, Captain, Crew
}
\examples{
multiplayer_game(4, c("greedy", "greedy"))
multiplayer_game(100, c("none", "greedy"))

}
\seealso{
Other simulation functions: \code{\link{reset_game}}
}
\concept{simulation functions}
