% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_helpers.R
\name{play_again}
\alias{play_again}
\title{play_again}
\usage{
play_again(score)
}
\arguments{
\item{score}{final player score of previous game}
}
\value{
TRUE if player will start another game, FALSE otherwise
}
\description{
Ask user if they would like to play another game
}
\seealso{
Other Interactive Gameplay: \code{\link{find_score}},
  \code{\link{game_turn}}, \code{\link{player_message}},
  \code{\link{reroll_prompt}}, \code{\link{reroll}},
  \code{\link{start_game}}
}
\concept{Interactive Gameplay}
