% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_helpers.R
\name{player_message}
\alias{player_message}
\title{player_message}
\usage{
player_message(components, score, num_rolls)
}
\arguments{
\item{components}{\itemize{
\item dice to keep from roll
}}

\item{score}{\itemize{
\item current player score
}}

\item{num_rolls}{\itemize{
\item number of turns a player has played
}}
}
\description{
Helper function for interactive gameplay
Print messages in response to player rolls
}
\seealso{
Other Interactive Gameplay: \code{\link{find_score}},
  \code{\link{game_turn}}, \code{\link{play_again}},
  \code{\link{reroll_prompt}}, \code{\link{reroll}},
  \code{\link{start_game}}
}
\concept{Interactive Gameplay}
