% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_helpers.R
\name{reroll}
\alias{reroll}
\title{reroll}
\usage{
reroll(num_dice)
}
\arguments{
\item{num_dice}{number of dice to be rolled}
}
\value{
vector of dice rolls equal to num_dice
}
\description{
Dice rolling function; Simulates a number of dice rolls between 1 and 5
}
\seealso{
Other Interactive Gameplay: \code{\link{find_score}},
  \code{\link{game_turn}}, \code{\link{play_again}},
  \code{\link{player_message}},
  \code{\link{reroll_prompt}}, \code{\link{start_game}}
}
\concept{Interactive Gameplay}
