% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_helpers.R
\name{reroll_prompt}
\alias{reroll_prompt}
\title{reroll_prompt}
\usage{
reroll_prompt(num_arg)
}
\arguments{
\item{num_arg}{current turn number of player; must be between 1 and 3}
}
\value{
TRUE if player will reroll, FALSE otherwise
}
\description{
Ask user if they would like to roll again and collects player answer
}
\seealso{
Other Interactive Gameplay: \code{\link{find_score}},
  \code{\link{game_turn}}, \code{\link{play_again}},
  \code{\link{player_message}}, \code{\link{reroll}},
  \code{\link{start_game}}
}
\concept{Interactive Gameplay}
