% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{reset_game}
\alias{reset_game}
\title{reset_game}
\usage{
reset_game(...)
}
\arguments{
\item{None}{}
}
\description{
Configures global environment to allow for repeated game simulations
}
\examples{
reset_game()

}
\seealso{
Other simulation functions: 
\code{\link{multiplayer_simulation}()}
}
\concept{simulation functions}
