% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{reset_simulation}
\alias{reset_simulation}
\title{reset_simulation}
\usage{
reset_simulation()
}
\description{
Configures global environment to allow for repeated game simulations
}
\seealso{
Other simulation functions: \code{\link{multiplayer_simulation}},
  \code{\link{start_simulation}}
}
\concept{simulation functions}
