% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{rolling_function}
\alias{rolling_function}
\title{rolling_function}
\usage{
rolling_function(player_dice, num_dice, num_games, num_rolls, strategy,
  game_ticker, gamecard, scores_vector, multiplayer)
}
\arguments{
\item{player_dice}{a vector of the dice saved to a player's hand in a given game}

\item{num_dice}{the number of dice a player has left to roll}

\item{num_games}{the number of games a user wishes to run within the simulation}

\item{num_rolls}{the number of rolls a player has left to use}

\item{strategy}{a character vector containing p1 and p2 strategies}

\item{game_ticker}{the number of games already completed in the simulation}

\item{gamecard}{a dataframe storing game number and score information}

\item{scores_vector}{a vector of a player's scores}

\item{multiplayer}{a logical vector determining whether games are single or multiplayer}
}
\description{
Simulates rolling of a player's remaining dice and calls initial_gameplay
}
\seealso{
Other helper functions: \code{\link{final_score_tab}},
  \code{\link{greedy_reroll}},
  \code{\link{initial_gameplay}},
  \code{\link{strategy_function}}
}
\concept{helper functions}
