% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{start_game}
\alias{start_game}
\title{Initializes gameplay for a single player simulation of Ship, Captain, Crew}
\usage{
start_game(...)
}
\arguments{
\item{num_games}{the number of games to be simulated}

\item{strategy}{a character vector containing the player's strategy}
}
\description{
Initializes gameplay for a single player simulation of Ship, Captain, Crew
}
\examples{
start_game(10, "greedy")

}
