% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{start_simulation}
\alias{start_simulation}
\title{Initializes gameplay for a single player simulation of Ship, Captain, Crew}
\usage{
start_simulation(num_games, strategy = "default", multiplayer = FALSE)
}
\arguments{
\item{num_games}{the number of games to be simulated}

\item{strategy}{a character vector containing the player's strategy}
}
\description{
Initializes gameplay for a single player simulation of Ship, Captain, Crew
}
\examples{
\dontrun{start_simulation(10, "greedy")}

}
\seealso{
Other simulation functions: \code{\link{multiplayer_simulation}},
  \code{\link{reset_simulation}}
}
\concept{simulation functions}
