% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_functions.R
\name{start_game}
\alias{start_game}
\title{start_game}
\usage{
start_game(...)
}
\arguments{
\item{...}{user-provided additional arguments, not needed for regular gameplay}
}
\description{
User-playable ship, captain, crew game
}
\examples{
\dontrun{start_game()}
}
\seealso{
Other Interactive Gameplay: \code{\link{find_score}},
  \code{\link{game_turn}}, \code{\link{play_again}},
  \code{\link{player_message}},
  \code{\link{reroll_prompt}}, \code{\link{reroll}}
}
\concept{Interactive Gameplay}
